Создал отдельный каталог  для хранения  перечислений
