+++ TrackerVC/Habit/NewHabitViewController.swift
@@ -418,12 +418,16 @@ extension NewHabitViewController: UITableViewDataSource, UITableViewDelegate {
 extension NewHabitViewController: UITextFieldDelegate {
     
     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        
         guard let textFieldText = textField.text else { return true }
+        
         let newText = (textFieldText as NSString).replacingCharacters(in: range, with: string)
         let count = newText.count < 38 ? newText.count : 38
         let remainingCharacters = 38 - count
+        
         subtitleNameTracker.text = count < 38 ? ("Осталось \(remainingCharacters) символов") : ("Ограничение \(count) символов")
         subtitleNameTracker.textColor = count < 38 ? .ypLightGray : .ypRed
+        
         return newText.count <= 38
     }
